Created Project - Extra
In this verson i created the project and the test project. Cottrell did say that we could all use the same tests, but we had to have some unique test. My Idea is that we can all make test for the app, but for what you will be desiging specifically, id imageine we will have a good some of 10-15 ish test that we all submit, and like 3- 5  unique tests in our submition. Be sure to label your Tests with like comments so when it comes time to get unique test you can easyly find yours.